+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Report index and shard implementations
 - Pagination support
 - Boosted index implementation
+- Topic index implementation